### 2019-10-18 #1820 @whedon
[REVIEW]: memochange: An R package for estimation procedures and tests for persistent time series
**Submitting author:** @KaiWenger (<a href="http://orcid.org/0000-0002-6566-0427">Kai Wenger</a>)
**Repository:** <a href="https://github.com/KaiWenger/memochange" target ="_blank">https://github.com/KaiWenger/memochange</a>
**Version:** 1.1.0
