### 2019-10-22 9f594f1
Corrected some typos
