### 2019-10-29 KaiWenger/memochange#1 @steffilazerte
JOSS Review: Installation/Loading suggestions
Hi @KaiWenger, Hi @Janis410!

I'm one of your reviewers for the JOSS submission (https://github.com/openjournals/joss-reviews/issues/1820). I'm going to be opening Issues here as part of my review process.

---
- [x] Imported packages should have minimum versions (it'll save headaches later on!)
e.g., in the `DESCRIPTION` file: `Imports: LongMemoryTS (>= 0.1.0), ...`
Use `packageVersion("LongMemoryTS")` to find out which package version you've been testing with.
- [x] Consider adding `.travis.yml`, `cran-comments.md`, `figures`, `paper` to `.Rbuildignore` (will ensure you, or other people forking and building your package, don't get CHECK notes later on)
- [x] Only import the functions you use (or, even better, use the `package::function()` notation). See below

There aren't that many places where you use functions from other packages. In these situations, it's generally considered better practice to call the function directly rather than to load the whole package; Your package will load more quickly and will be less likely to conflict with other packages.

For example, in `BP_estim()`, consider using `stats::lm()`, `fracdiff::fdBPH()`, and `LongMemoryTS::ELS()`. If you did this everywhere, you could then remove the `add_to_Namespace.R` file (but leave the packages under Imports in the `DESCRIPTION` file)

At the very least, I would consider only importing the functions you use, rather than the entire package (e.g., `@importFrom`).

Reference: <http://r-pkgs.had.co.nz/namespace.html#imports>
