### 2019-10-29 KaiWenger/memochange#2 @steffilazerte
JOSS Review: README suggestions
Hi @KaiWenger, Hi @Janis410!

I'm one of your reviewers for the JOSS submission (https://github.com/openjournals/joss-reviews/issues/1820). I'm going to be opening Issues here as part of my review process.

---
- [x] Consider including .Rmd file and compiling this to create the github .md file (means that when you update the package you just have to re-compile the Rmd file rather than copy/paste results by hand)
- [x] The `README` examples are very useful! I found a couple of places that could use some clarification/correction:
  - after the `ratio_test` example, what is the "break fraction"? tau?
  - a section compiled with "$" looks like latex code which wasn't correctly compiled. You may need to check and see if LATEX code compiles in md files for GitHub
- [x] Consider including the output of all the examples. It gets long and a bit messy, but will help avoid confusing users
  - E.g., `BP_estim(x, direction = "01")` etc.
- [x] I got lost with the change-in-mean example. I would suggest adding a simpler example before the complex comparison one.
