### 2019-10-29 KaiWenger/memochange#3 @steffilazerte
JOSS Review: Documentation suggestions
Hi @KaiWenger, Hi @Janis410!

I'm one of your reviewers for the JOSS submission (https://github.com/openjournals/joss-reviews/issues/1820). I'm going to be opening Issues here as part of my review process.

---
**General documentation**

- [x] Consider including vignettes or demo (e.g., the `README` from the github site, could be adapted and included as vignette).
- [x] Consider fleshing out the general description in the documentation for `memochange` (i.e. that in `memochange.R`). So users get a more detailed overview when they use `?memochange`.

**Examples**
Including examples everywhere is great. However, for some examples, it isn't clear what part of the example is about creating dummy data, and what part is about running the test.

- [x] Consider adding comments to clarify things for the user.

E.g.,  for `CUSUMLM()` (assuming I've understood what the different parts are for)

```
library(fracdiff) # To create dummy data
library(LongMemoryTS) # To determine the d_est parameter

# Create dummy data set to test
n        <- 500
d        <- 0.2
tseries  <- fracdiff.sim(n=n,d=d)$series

# Determine the d_est parameter
d_est    <- local.W(tseries, m=floor(1+n^0.65))$d

# Create 2nd dummy data set to test against
changep  <- c(rep(0,n/2),rep(1,n/2))
tseries2 <- tseries+changep

# Determine the d_est parameter
d_est2   <- local.W(tseries2, m=floor(1+n^0.65))$d

# Test for a change-in-mean
CUSUMLM(tseries,delta=0.65,d=d_est)
CUSUMLM(tseries2,delta=0.65,d=d_est2)
```

**Unit Tests**

All the functions do include examples which are run and tested by travisCI, but this only tests for the occurrence of errors, not for expected output.
- [x] I would HIGHLY recommend including `testthat` tests and even consider adding a coverage metric (proportion of your code that is tested). You can automatically set this up with `usethis::use_coverage()`

**Community Guidelines**

There is a statement at the end of the `README`, but it isn't clear if software contributions are welcome or not.

- [x] Consider including a `CONTRIBUTING.md` file (e.g., https://github.com/ropensci/weathercan/blob/master/CONTRIBUTING.md)
