### 2019-10-29 #1820 @steffilazerte
Hi @KaiWenger, Hi @Janis410!

I have completed my first round of reviews and have some comments/suggestions for you. To make things clear, I've included a check box where I need feedback or action from you. This R package is solid and and well documented. I found the review straightforward and the examples easy to follow. Thanks!

**Authorship**
- [ ] Could you supply me with a brief statement of author contributions? This is not strictly required by JOSS, but I find it helps me evaluate the "Contribution and authorship" requirements

I installed the CRAN version. I had to install some other libraries first, but this kind of thing is common for my OS (Linux) and was easily fixed. The installation went smoothly and I ran into no problems testing the package.
- [ ] I have included some suggestions for "good-practices" (KaiWenger/memochange#1)

I was thoroughly impressed by the amount of documentation that has gone into the functions, especially with respect to referencing different techniques. This is an area that is often skipped or only given lip service and the level of detail here makes the `memochange` package so much more usable!

- [ ] I have some suggestions for adding detail to the README (KaiWenger/memochange#2)
- [ ] I also suggest creating a vignette, adding detail to `memochange.R` and annotating the examples  (KaiWenger/memochange#3)
- [ ] The package examples are tested by TravisCI, and I suggest adding unit tests and potentially test coverage (KaiWenger/memochange#3)
- [ ] I think explicit community guidelines would be helpful (KaiWenger/memochange#3)

I am a biologist and R programmer familiar with statistical analyses in general, but not with time series analyses nor economics/finance. I was pleasantly surprised to find that while I found the summary complex, it was still mostly understandable (although don't ask me to explain any of the details!). However, it is a complex write up and I'm curious what @JonasMoss thought.

- [ ] Are there any other packages that do similar things?
- [ ] While well written, I do have some minor editorial comments:
  - Page 1, Start of paragraph 5 - "In recent years, **many** researchers"
  - Page 2, L1 - "adapted **to** the CUSUM, etc."  (?)
  - Page 2, L1/2 - "or Wilcoxon test**s**."   (?)
  - Page 2, L6/7 - "This may be **because** Wenger et al. (2019) **found** that ..."
  - Page 2 Statement of Need: put the package name "memochange" in `code font` (i.e., use the ticks `)
  - Page 2, Statement of Need L6/7 : "there is a wide range of possible time series **to which the procedures can be applied**, eg. ..."
- [ ] Include link to github repository
- [ ] Citations in brackets should go in square braces to ensure the correct formatting e.g., `[see @robinson1995gaussian, ...]` (I think this should remove the double brackets around citations)

Don't hesitate to let me know if anything is unclear or you would like more details.

---
**DISCLAIMER:** @mbobra  I know nothing of the particular analyses being performed here. My review consists of reviewing whether the package runs into errors, has documentation, examples, and compiles correctly. I am able to reproduce the examples, but am unable to assess whether the results returned are, in fact, *statistically* sound/correct.
