### 2019-10-29 KaiWenger/memochange#5 @JonasMoss
JOSS Review: Be explicit with dependencies in examples.
It would be nice not to use  `library(lib)` in the examples, but call the functions with explicit reference to the library it is imported from, e.g
```
# Don't do:
library(fracdiff)
tseries  <- fracdiff.sim(n=n,d=d)$series

# Do this:
tseries  <- fracdiff::fracdiff.sim(n=n,d=d)$series
```
And please remove redundant `library` calls, such `library(LongMemoryTS)` in the example of `pb_sim`.
```
#' library(LongMemoryTS)
#' series<- pb_sim(500,0.5,"none",d1=0.2,d2=0.8,mean=0,var=1)
#' ts.plot(series)
```
This is related to https://github.com/KaiWenger/memochange/issues/4. The main reason to do this is to aid your reader understanding what's going on.
