### 2019-10-29 KaiWenger/memochange#6 @JonasMoss
JOSS Review: Add tests
This package has no automated tests except for the examples. The easiest way to do testing in R is the `testthat` package. You need to implement automatic tests for this package, as I need to trust that all the exported functions do what they should.

1. Need tests of most reasonable combinations of use cases, e.g. types, bandwidths, small time series, large time series.
2. Should aim for a pretty high coverage.

This is the most important fix, with the dependency problem https://github.com/KaiWenger/memochange/issues/4 coming in second.

What's more, would it be possible to demonstrate your functions do what they should by replicating computations from papers?
