### 2019-10-29 KaiWenger/memochange#8 @JonasMoss
JOSS Review: Specific function comments
Here are comments on some specific functions. I have not gone through all in similar detail and maybe the others have similar issues.

### MR_test

#### Strong Suggestions
* For the `trend` component, write `\code{"linear"}` instead of `\code{linear}`.  The error messages have the same problem.
* Fix the same issue with `statistics`. Write e.g. `"standard"` instead of `standard`. The error messages have the same problem.
* Avoid `if (mode(x) %in% ("numeric") == FALSE | is.vector(x)==FALSE)
    stop("x must be a univariate numeric time series")` as "univariate numeric time series" isn't an object. You want `x` to be a *numeric vector*. It is interpreted as a time series but isn't a time series *object*.
* Likewise, you should write in the documentation that `x` cannot contain missing values when you write `if (any(is.na(x)))
    stop("missing values not allowed in time series")` Consider writing rewriting the error message to e.g. "x contains missing values".
* This demand should also be documented:`if ((T*tau)<=2+ as.numeric(p>1)*(p))
    stop("increase T*tau to guarantee that the test statistic can be calculated")`

#### Weak Suggestions
* Remove `library(fracdiff)`and `library(LongMemoryTS)` from the examples; they are not used.
* I don't know how to interpret the results.

#### Very Weak Suggestions
* Check that `trend` is in `c("none","linear")` using `trend = match.arg(trend)` in the body of the code. You can use `match.arg` in the same way with your other checks.
* Don't write  `trend<-trend[1]`, `statistic<-statistic[1]` etc. This is for handling defaults, but is automatically done using `match.arg` above.
* You state "If \code{simu=1}, the critical values are simulated based on the given data using M replications. Caution, for large M this might take a while." Please say what a large `M` is and what it "takes a while" means.
* Regardining "The critical values of the tests vary with sample size and memory parameter d. If \code{simu=0}, the critical values provided are based on linear interpolation of the critical values simulated by Martins and Rodrigues (2014). These are based on \code{tau=0.2}." Consider adding a note when `simu=0` but `tau!=0.2`.

### pb_sim
#### Strong Suggestions
* Document exactly what kind of time series is generated. It seems to be a fractional white noise process to me.
* For the `trend` component, write `\code{"linear"}` instead of `\code{linear}` and the same for `\code{none}`. Writing `pb_sim(500,0.5,linear,d1=0.2,d2=0.8,mean=0,var=1)` throws an error when `linear` is not defined.
#### Weak Suggestions
* Explicitly state you are simulating two times from `
LongMemoryTS::FI.sim` in the documentation.
* Please check that `trend` is in `c("none","linear")` using e.g. `trend = match.arg(trend)` in the body of the code. Now `pb_sim(500,0.5,"limear",d1=0.2,d2=0.8,mean=0,var=1)` returns a timeseries without a trend but should throw an error. Such functions are bug prone.
#### Suggestions
* It's wise to avoid `T` as an a variable name, as `T` is an abbreviation for `TRUE` in `R`.

### wilcoxonLM

#### Weak Suggestions
* Example: Remove library calls.
* Example: Comment on why you use `fracdiff::fracdiff.sim` and what it is.
* Example: Add `set.seed` and comment on the results. Right now your intended results are not reproducible. It would be nice to have some proof that the procedure works as intended in a vignette using e.g. Monte Carlo.
* Example: Comment on what `local.W(tseries2, m=floor(1+n^0.65))$d` is and why you use ` m=floor(1+n^0.65)`. Don't forget to add the `::LongMemoryTS` moniker in front of it.
(This is in line with the suggestions in https://github.com/KaiWenger/memochange/issues/3 )
