### 2019-10-29 KaiWenger/memochange#9 @JonasMoss
JOSS Review: Readme and vignette
Here are some suggestions for the readme and vignettes.

First, please add a description to the Github page!

## Readme and vignette
The example usage is too long. I think your readme should be turned into a vignette or even two vignettes instead. An example usage should ideally be short and to the point, using little code and not much text.

The package contains *many* fuctions but no high-level overview of how they are related or why they are written. It would be great to have this, in the form of a vignette giving an overview of the package, perhaps? Now I have looked at this package for ~5 hours and I only have a small idea of what they do or how they are related. Also, see the first point of https://github.com/KaiWenger/memochange/issues/3

The long example in the readme is a step in the right direction for solving this problem, but this should ideally be a vignette as I said above, and there it should probably be split into two vignettes too. The package appears to do two things (in many ways!): Checking for breakpoints in persistence *and* checking for a change in mean. It would be nice to have vignettes covering both, as they are separate problems. Maybe more than two too, as there are *many* top-level functions to cover.
