+++ README.md
@@ -41,8 +41,7 @@ In this section we will show how the implemented procedures can be used on a rea
 As an empirical example for a process that might exhibit a break in persistence, we consider the price of crude oil.  First, we download the monthly price series from the FRED data base. For this purpose we need the `data.table` package.
 
 ```r
-library(data.table)
-oil=fread("https://fred.stlouisfed.org/graph/fredgraph.csv?bgcolor=%23e1e9f0&chart_type=line&drp=0&fo=open%20sans&graph_bgcolor=%23ffffff&height=450&mode=fred&recession_bars=on&txtcolor=%23444444&ts=12&tts=12&width=1168&nt=0&thu=0&trc=0&show_legend=yes&show_axis_titles=yes&show_tooltip=yes&id=MCOILWTICO&scale=left&cosd=1986-01-01&coed=2019-08-01&line_color=%234572a7&link_values=false&line_style=solid&mark_type=none&mw=3&lw=2&ost=-99999&oet=99999&mma=0&fml=a&fq=Monthly&fam=avg&fgst=lin&fgsnd=2009-06-01&line_index=1&transformation=lin&vintage_date=2019-09-23&revision_date=2019-09-23&nd=1986-01-01")
+oil=data.table::fread("https://fred.stlouisfed.org/graph/fredgraph.csv?bgcolor=%23e1e9f0&chart_type=line&drp=0&fo=open%20sans&graph_bgcolor=%23ffffff&height=450&mode=fred&recession_bars=on&txtcolor=%23444444&ts=12&tts=12&width=1168&nt=0&thu=0&trc=0&show_legend=yes&show_axis_titles=yes&show_tooltip=yes&id=MCOILWTICO&scale=left&cosd=1986-01-01&coed=2019-08-01&line_color=%234572a7&link_values=false&line_style=solid&mark_type=none&mw=3&lw=2&ost=-99999&oet=99999&mma=0&fml=a&fq=Monthly&fam=avg&fgst=lin&fgsnd=2009-06-01&line_index=1&transformation=lin&vintage_date=2019-09-23&revision_date=2019-09-23&nd=1986-01-01")
 ```
 
 To get a first visual impression, we plot the series.
@@ -51,13 +50,13 @@ To get a first visual impression, we plot the series.
 library(zoo)
 library(xts)
 oil=as.data.frame(oil)
-oil$DATE=as.Date(oil$DATE)
-oil_xts=xts(oil[,-1],order.by = oil$DATE)
-plot.zoo(oil_xts,xlab="",ylab="Price",main="Crude Oil Price: West Texas Intermediate")
+oil$DATE=zoo::as.Date(oil$DATE)
+oil_xts=xts::xts(oil[,-1],order.by = oil$DATE)
+zoo::plot.zoo(oil_xts,xlab="",ylab="Price",main="Crude Oil Price: West Texas Intermediate")
 ```
 ![](figures/oil_plot.png)
 
-From the plot we observe that the series seems to be more variable in its second part from year 2000 onwards.  This is first evidence that a change in persistence has occurred. We can test this hypothesis using the functions `cusum_test`, `LBI_test`, `LKSN_test`, `MR_test`, and `ratio_test`. In this short example we use the ratio and MR test since these are the empirically most often applied one. The functionality of the other tests is similar. They all require a univariate numeric time series `x` as an input variable and yield a matrix of test statistic and critical values as an output variable.
[…]
