### 2019-10-30 d27062d
Delete add_to_Namespace.R
not needed anymore
