+++ R/fixbsupw.R
@@ -4,47 +4,64 @@
 #' The function returns the test statistic as well as critical values.
 #' @details
 #' Note that the critical values are generated for \code{tau=0.15} using the Bartlett kernel.
-#' @param tseries the univariate numeric time series to be investigated.
-#' @param d the long-memory parameter.
-#' @param bandw the bandwidth parameter for the long-run variance estimator. It can take values in the range \code{bandw=[0.05,0.1,0.2]}. Default is
+#' @param x the univariate numeric vector to be investigated. Missing values are not allowed.
+#' @param d integer that specifies the long-memory parameter.
+#' @param bandw integer that determines the bandwidth parameter for the long-run variance estimator. It can take values in the range \code{bandw=[0.05,0.1,0.2]}. Default is
 #' \code{bandw=0.1}, which is suggested by Iacone, Leybourne and Taylor (2014).
-#' @param tau defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).
-#' @return Returns the test statistic and the corresponding critical values of the test.
+#' @param tau integer that defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).
+#' @return Returns a numeric vector containing the test statistic and the corresponding critical values of the test.
+#' @seealso \code{\link{CUSUMfixed}}, \code{\link{snsupwald}}
 #' @author Kai Wenger
 #' @examples
-#' library(fracdiff)
-#' library(LongMemoryTS)
-#'
-#' n        <- 500
+#' # set model parameters
+#' T        <- 500
 #' d        <- 0.2
-#' tseries  <- fracdiff.sim(n=n,d=d)$series
-#' d_est    <- local.W(tseries, m=floor(1+n^0.65))$d
+#' 
+#' set.seed(410)
+#' 
+#' # generate a fractionally integrated (long-memory) time series
+#' tseries  <- fracdiff::fracdiff.sim(n=T, d=d)$series
 #'
-#' changep  <- c(rep(0,n/2),rep(1,n/2))
+#' # generate a fractionally integrated (long-memory) time series with a change in mean in the middle of the series
+#' changep  <- c(rep(0,T/2), rep(1,T/2))
 #' tseries2 <- tseries+changep
-#' d_est2   <- local.W(tseries2, m=floor(1+n^0.65))$d
+#' 
+#' # estimate the long-memory parameter of both series via local Whittle approach. The bandwidth to estimate d is chosen as T^0.65, which is usual in literature
+#' d_est    <- LongMemoryTS::local.W(tseries, m=floor(1+T^0.65))$d
+#' d_est2   <- LongMemoryTS::local.W(tseries2, m=floor(1+T^0.65))$d
 #'
-#' fixbsupw(tseries,d=d_est)
-#' fixbsupw(tseries2,d=d_est2)
+#' # perform the test on both time series
+#' fixbsupw(tseries, d=d_est)
+#' fixbsupw(tseries2, d=d_est2)
+#' # For the series with no change in mean the test does not reject the null hypothesis of a constant mean across time at any reasonable significance level.
+#' # For the series with a change in mean the test rejects the null hypothesis at a 1% significance level.
 #' @references
[…]
