+++ R/ratio_test.R
@@ -14,24 +14,30 @@
 #' It should be noted, however, that only the critical values are identical, the distribution is highly irregular.
 #'
 #' The critical values of the tests vary with the sample size. If \code{simu=0}, the critical values provided
-#' are based on linear interpolation of the critical values simulated by Harvey, Leybourne, and Taylor (2006). These are based on \code{tau=0.2}. If \code{simu=1},
-#' the critical values are simulated based on the given setup using M replications. Caution, for large M this might take a while,
-#' small M, however, make the results unreliable.
-#'
-#' @param x the univariate numeric time series to be investigated.
-#' @param trend whether the time series exhibits a trend, \code{none} implies no trend and \code{linear} implies a linear trend.
-#' @param tau the function tests in the interval \code{[T*tau,T*(1-tau)]} for a break in persistence with T being the length of the time series. It must hold that \code{0<tau<0.5}, default is \code{tau=0.2} as commonly used in the literature.
-#' @param type which type of ratio test should be performed, \code{BT} for the standard ratio test by Busetti and Taylor (2004), \code{LT} for the modified ratio test by Leybourne and Taylor (2004), and \code{HLT} respectively \code{HLTmin} are the modified tests by Harvey, Leybourne, and Taylor (2006). See details.
-#' @param statistic which type of test statistic should be used, \code{mean} corresponds to Hansen's (1991) mean score, \code{max} to Andrews' (1993) maximum statistic, and \code{exp} to Andrews and Ploberger's (1994) mean-exponential statistic
-#' @param m Number of covariances used for the estimation of the long run variance if \code{LT} is considered. Default is \code{m=0}.
-#' @param z Number of polynomials used if \code{HLT} or \code{HLTmin} are considered. Default is \code{z=9}.
+#' are based on linear interpolation of the critical values simulated by Harvey, Leybourne, and Taylor (2006). These are, however, only valid for \code{tau=0.2}, \code{m=0}, and \code{z=9}. 
+#' In case that non-default values are chosen for \code{tau}, \code{m}, or \code{z}, it is recommended to set \code{simu=1} which means that critical values are simulated based on the given data using M replications. Caution, for M=10,000 this takes at least thirty minutes, depending on the length of the time series.
+#' Smaller M, however, make the results unreliable.
+#' 
+#' @param x the univariate numeric vector to be investigated. Missing values are not allowed.
+#' @param trend whether the time series exhibits a trend, \code{"none"} implies no trend and \code{"linear"} implies a linear trend.
[…]
