+++ R/CUSUMLM.R
@@ -4,28 +4,37 @@
 #' The function returns the test statistic as well as critical values.
 #' @details
 #' Note that the critical values are generated for \code{tau=0.15}.
-#' @param tseries the univariate numeric time series to be investigated.
-#' @param d the long-memory parameter.
-#' @param delta the bandwidth that is used to estimate the constant \code{G} that approximates the short run dynamics of the time series at the origin.
+#' @param x the univariate numeric vector to be investigated. Missing values are not allowed.
+#' @param d integer that specifies the long-memory parameter.
+#' @param delta integer that determines the bandwidth that is used to estimate the constant \code{G} that approximates the short run dynamics of the time series at the origin.
 #' The same bandwidth should be used that is applied to estimate \code{d} before. See Wenger, Leschinski, Sibbertsen (2018) for details.
-#' @param tau defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).
-#' @return Returns the test statistic and the corresponding critical values of the test.
+#' @param tau integer that defines the search area, which is \code{[tau,1-tau]}. Default is \code{tau=0.15} as suggested by Andrews (1993).
+#' @return Returns a numeric vector containing the test statistic and the corresponding critical values of the test.
+#' @seealso \code{\link{CUSUMfixed}}, \code{\link{CUSUM_simple}}
 #' @author Kai Wenger
 #' @examples
-#' library(fracdiff)
-#' library(LongMemoryTS)
-#'
-#' n        <- 500
+#' # set model parameters
+#' T        <- 500
 #' d        <- 0.2
-#' tseries  <- fracdiff.sim(n=n,d=d)$series
-#' d_est    <- local.W(tseries, m=floor(1+n^0.65))$d
+#' 
+#' set.seed(410)
+#' 
+#' # generate a fractionally integrated (long-memory) time series
+#' tseries  <- fracdiff::fracdiff.sim(n=T, d=d)$series
 #'
-#' changep  <- c(rep(0,n/2),rep(1,n/2))
+#' # generate a fractionally integrated (long-memory) time series with a change in mean in the middle of the series
+#' changep  <- c(rep(0,T/2), rep(1,T/2))
 #' tseries2 <- tseries+changep
-#' d_est2   <- local.W(tseries2, m=floor(1+n^0.65))$d
+#' 
+#' # estimate the long-memory parameter of both series via local Whittle approach. The bandwidth to estimate d is chosen as T^0.65, which is usual in literature
+#' d_est    <- LongMemoryTS::local.W(tseries, m=floor(1+T^0.65))$d
+#' d_est2   <- LongMemoryTS::local.W(tseries2, m=floor(1+T^0.65))$d
 #'
-#' CUSUMLM(tseries,delta=0.65,d=d_est)
-#' CUSUMLM(tseries2,delta=0.65,d=d_est2)
+#' # perform the test on both time series
+#' CUSUMLM(tseries, delta=0.65, d=d_est)
+#' CUSUMLM(tseries2, delta=0.65, d=d_est2)
+#' # For the series with no change in mean the test does not reject the null hypothesis of a constant mean across time at any reasonable significance level.
+#' # For the series with a change in mean the test rejects the null hypothesis at a 1% significance level.
 #' @references
[…]

[…]
