+++ man/pb_sim.Rd
@@ -11,7 +11,7 @@ pb_sim(T, tau, trend = c("none", "linear"), tp = 1, d1, d2, mean, var)
 
 \item{tau}{break fraction, \code{T*tau} yields the break point. It needs to hold that \code{0<tau<1}.}
 
-\item{trend}{whether the time series exhibits a trend, \code{none} implies no trend and \code{linear} implies a linear trend.}
+\item{trend}{whether the time series exhibits a trend, \code{"none"} implies no trend and \code{"linear"} implies a linear trend.}
 
 \item{tp}{trend parameter, \code{t*tp} yields the contribution of the trend component if \code{trend="linear"}.}
 
@@ -27,12 +27,16 @@ pb_sim(T, tau, trend = c("none", "linear"), tp = 1, d1, d2, mean, var)
 Returns a vector containing the simulated time series.
 }
 \description{
-This function simulates a process that exhibits a break in persistence.
+This function simulates a fractional white noise process that exhibits a break in persistence using \code{FI.sim} from the \code{LongMemoryTS} package. In the first part of the series the noise is integrated with order \code{d_1} and in the second part with order \code{d_2}.
 }
 \examples{
-library(LongMemoryTS)
-series<- pb_sim(500,0.5,"none",d1=0.2,d2=0.8,mean=0,var=1)
-ts.plot(series)
+set.seed(410)
+
+# generate persistence-break time series
+series <- pb_sim(500, 0.5, "none", d1=0.2, d2=0.8, mean=0, var=1)
+
+# plot generated series
+stats::ts.plot(series)
 }
 \author{
 Janis Becker
