+++ README.Rmd
@@ -1,0 +1,180 @@
+---
+bibliography: paper/paper.bib
+biblio-style: "apalike"
+output:
+  github_document:
+    html_preview: true
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# memochange
+## Testing for Structural Breaks under Long Memory and Testing for Changes in Persistence
+[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/memochange)](https://CRAN.R-project.org/package=memochange)
+[![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange)
+[![License: GPL v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0) 
+[![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+
+
+The `memochange` package implements the most prominent test procedures and break point estimators for persistent processes that exhibit structural breaks in mean or in persistence.
+On the one hand, the package contains the most popular approaches for testing whether a time series exhibits a break in persistence from I(0) to I(1) or vice versa. In case the tests reject the null of constant persistence, various breakpoint estimators are available to detect the point of the break as well as the order of integration in the two regimes.
+On the other hand, the package contains the most popular approaches to test for a change-in-mean in a long-memory time series. 
+These include memory robust versions of the CUSUM, sup-Wald, and Wilcoxon type tests. The tests either utilize consistent estimates of the long-run variance or a self normalization approach in their test statistics.
+
+`memochange` may be of interest to those who analyse macroeconomic and financial time series such as inflation rates, trading volume, interest rates, volatilities and so on.
+
+
+You can install this R package from GitHub:
+
+
+```r
+install.packages("devtools")
+library(devtools)
+install_github("KaiWenger/memochange")
+```
+
+or directly from the CRAN repository:
+
+
+```r
+install.packages("memochange")
+```
+
+
+## Examples
+
+In this section we present two short examples that illustrate how the implemented procedures can be used on a real data. A more detailed presentation of the various tests and functions can be found in the vignette.
+
+### Tests for persistence change
+
+As an empirical example for a process that might exhibit a break in persistence, we consider the price of crude oil.  First, we download the monthly price series from the FRED data base. For this purpose we need the `data.table` package.
+
+```r
+oil=data.table::fread("https://fred.stlouisfed.org/graph/fredgraph.csv?bgcolor=%23e1e9f0&chart_type=line&drp=0&fo=open%20sans&graph_bgcolor=%23ffffff&height=450&mode=fred&recession_bars=on&txtcolor=%23444444&ts=12&tts=12&width=1168&nt=0&thu=0&trc=0&show_legend=yes&show_axis_titles=yes&show_tooltip=yes&id=MCOILWTICO&scale=left&cosd=1986-01-01&coed=2019-08-01&line_color=%234572a7&link_values=false&line_style=solid&mark_type=none&mw=3&lw=2&ost=-99999&oet=99999&mma=0&fml=a&fq=Monthly&fam=avg&fgst=lin&fgsnd=2009-06-01&line_index=1&transformation=lin&vintage_date=2019-09-23&revision_date=2019-09-23&nd=1986-01-01")
+```
+
+To get a first visual impression, we plot the series.
+
+```{r, echo = TRUE, fig.height = 4, fig.width = 7, fig.align = "center"}
+oil=as.data.frame(oil)
[…]
