+++ README.md
@@ -1,24 +1,20 @@
 
-# memochange
-## Testing for Structural Breaks under Long Memory and Testing for Changes in Persistence
-[![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/memochange)](https://CRAN.R-project.org/package=memochange)
-[![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange)
-[![License: GPL v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0) 
-[![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+memochange
+==========
 
+Testing for Structural Breaks under Long Memory and Testing for Changes in Persistence
+--------------------------------------------------------------------------------------
 
-The `memochange` package implements the most prominent test procedures and break point estimators for persistent processes that exhibit structural breaks in mean or in persistence.
-On the one hand, the package contains the most popular approaches for testing whether a time series exhibits a break in persistence from I(0) to I(1) or vice versa. In case the tests reject the null of constant persistence, various breakpoint estimators are available to detect the point of the break as well as the order of integration in the two regimes.
-On the other hand, the package contains the most popular approaches to test for a change-in-mean in a long-memory time series. 
-These include memory robust versions of the CUSUM, sup-Wald, and Wilcoxon type tests. The tests either utilize consistent estimates of the long-run variance or a self normalization approach in their test statistics.
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/memochange)](https://CRAN.R-project.org/package=memochange) [![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange) [![License: GPL v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0) [![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+
+The `memochange` package implements the most prominent test procedures and break point estimators for persistent processes that exhibit structural breaks in mean or in persistence. On the one hand, the package contains the most popular approaches for testing whether a time series exhibits a break in persistence from I(0) to I(1) or vice versa. In case the tests reject the null of constant persistence, various breakpoint estimators are available to detect the point of the break as well as the order of integration in the two regimes. On the other hand, the package contains the most popular approaches to test for a change-in-mean in a long-memory time series. These include memory robust versions of the CUSUM, sup-Wald, and Wilcoxon type tests. The tests either utilize consistent estimates of the long-run variance or a self normalization approach in their test statistics.
 
 `memochange` may be of interest to those who analyse macroeconomic and financial time series such as inflation rates, trading volume, interest rates, volatilities and so on.
 
-
 You can install this R package from GitHub:
 
-
-```r
+``` r
 install.packages("devtools")
 library(devtools)
 install_github("KaiWenger/memochange")
@@ -26,243 +22,95 @@ install_github("KaiWenger/memochange")
 
 or directly from the CRAN repository:
 
-
-```r
+``` r
 install.packages("memochange")
 ```
 
+Examples
+--------
 
-## Examples
-
-In this section we will show how the implemented procedures can be used on a real data set as well as in Monte Carlo simulations to compare the performance of the tests.
[…]

[…]
