+++ vignettes/break_in_persistence.Rmd
@@ -1,0 +1,147 @@
+---
+title: "memochange-Tutorial: Break in Persistence"
+author: "Janis Becker"
+date: "`r Sys.Date()`"
+bibliography: paper.bib
+biblio-style: "apalike"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The `memochange` package can be used for two things: Checking for a break in persistence and checking for a change in mean. This vignette presents the functions related to a break in persistence. This includes `BP_estim`, `cusum_test`, `LBI_test`, `LKSN_test`, `MR_test`, `ratio_test`, and `pb_sim`. Before considering the usage of these functions, a brief literature review elaborates on their connection.
+
+##Literature Review
+The degree of memory is an important determinant of the characteristics of a time series. For an $I(0)$, or short-memory, process (e.g., AR(1) or ARMA(1,1)), the impact of shocks is short-lived and dies out quickly. On the other hand, for an $I(1)$, or difference-stationary, process such as the random walk, shocks persist infinitely. Thus, any change in a variable will have an impact on all future realizations. For an $I(d)$, or long-memory, process with $0<d<1$, shocks neither die out quickly nor persist infinitely, but have a hyperbolically decaying impact. In this case, the current value of a variable depends on past shocks, but the less so the further these shocks are past.
+
+There are plenty of procedures to determine the memory of a series (see @robinson1995gaussian, @shimotsu2010exact, among others). However, there is also the possibility that series exhibit a structural change in memory, often referred to as a change in persistence. Starting with @kim2000detection various procedures have been proposed to detect these changes and consistently estimate the change point. @busetti2004tests and @leybourne2004tests suggest approaches for testing the null of constant I(0) behaviour of the series against the alternative that a change from either $I(0)$ to $I(1)$ or $I(1)$ to $I(0)$ occurred. However, both approaches show serious distortions if neither the null nor the alternative is true, i.e. the series is constant $I(1)$. In this case the procedures by @leybourne2003tests and @leybourne2007cusum can be applied as they have the same alternative, but assume constant $I(1)$ behaviour under the null. Again, the procedures exhibit distortions when neither the null nor the alternative is true. To remedy this issue, @harvey2006modified suggest an approach that entails the same critical values for constant $I(0)$ and constant $I(1)$ behavior. Consequently, it accommodates both, constant $I(0)$ and constant $I(1)$ behavior under the null.
+
+While this earlier work focussed on the $I(0)/I(1)$ framework, more recent approaches are able to detect changes from $I(d_1)$ to $I(d_2)$ where $d_1$ and $d_2$ are allowed to be non-integers. @sibbertsen2009testing extend the approach of @leybourne2007cusum such that the testing procedure consistently detects changes from $0 \leq d_1<1/2$ to $1/2<d_2<3/2$ and vice versa. Under the null the test assumes constant $I(d)$ behavior with $0 \leq d <3/2$. The approach suggested by @martins2014testing is even able to identify changes from $-1/2<d_1<2$ to $-1/2<d_2<2$ with $d_1 \neq d_2$. Here, under the null the test assumes constant $I(d)$ behavior with $-1/2<d<2$.
+
[…]
