+++ testing/paper.bib
@@ -1,0 +1,326 @@
+@Manual{RCoreTeam,
+  title        = {R: A Language and Environment for Statistical
+                  Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = 2019,
+  url          = {https://www.R-project.org}
+}
+@Article{dickey1981likelihood,
+  Title={Likelihood ratio statistics for autoregressive time series with a unit root},
+  Author={Dickey, David A and Fuller, Wayne A},
+  Journal={Econometrica: journal of the Econometric Society},
+  Pages={1057--1072},
+  Year={1981},
+  doi={10.2307/1912517},
+  Publisher={JSTOR}
+}
+
+@Article{kwiatkowski1992testing,
+  Title={Testing the null hypothesis of stationarity against the alternative of a unit root: How sure are we that economic time series have a unit root?},
+  Author={Kwiatkowski, Denis and Phillips, Peter CB and Schmidt, Peter and Shin, Yongcheol},
+  Journal={Journal of econometrics},
+  Volume={54},
+  Number={1-3},
+  Pages={159--178},
+  Year={1992},
+  doi={10.1016/0304-4076(92)90104-Y},
+  Publisher={Elsevier}
+}
+
+@article{kim2000detection,
+  title={Detection of change in persistence of a linear time series},
+  author={Kim, Jae-Young},
+  journal={Journal of Econometrics},
+  volume={95},
+  number={1},
+  pages={97--116},
+  year={2000},
+  doi={10.1016/S0304-4076(99)00031-7},
+  publisher={Elsevier}
+}
+
+@article{busetti2004tests,
+  title={Tests of stationarity against a change in persistence},
+  author={Busetti, Fabio and Taylor, AM Robert},
+  journal={Journal of Econometrics},
+  volume={123},
+  number={1},
+  pages={33--66},
+  year={2004},
+  doi={10.1016/j.jeconom.2003.10.028},
+  publisher={Elsevier}
+}
+
+@article{leybourne2007cusum,
+  title={CUSUM of squares-based tests for a change in persistence},
+  author={Leybourne, Stephen and Taylor, Robert and Kim, Tae-Hwan},
+  journal={Journal of Time Series Analysis},
+  volume={28},
+  number={3},
+  pages={408--433},
+  year={2007},
+  doi={10.1111/j.1467-9892.2006.00517.x},
+  publisher={Wiley Online Library}
+}
+
+
+@article{leybourne2004tests,
+  title={On tests for changes in persistence},
+  author={Leybourne, Stephen and Taylor, AM Robert},
+  journal={Economics Letters},
+  volume={84},
+  number={1},
+  pages={107--115},
+  year={2004},
+  doi={10.1016/j.econlet.2003.12.015},
+  publisher={Elsevier}
+}
+
+
+@article{leybourne2003tests,
+  title={Tests for a change in persistence against the null of difference-stationarity},
+  author={Leybourne, Stephen and Kim, Tae-Hwan and Smith, Vanessa and Newbold, Paul},
+  journal={The Econometrics Journal},
+  volume={6},
+  number={2},
+  pages={291--311},
+  year={2003},
+  doi={10.1111/1368-423X.t01-1-00110},
+  publisher={Oxford University Press Oxford, UK}
+}
+
+@article{leybourne2007detecting,
+  title={Detecting multiple changes in persistence},
+  author={Leybourne, Stephen and Kim, Tae-Hwan and Taylor, AM Robert},
+  journal={Studies in Nonlinear Dynamics \& Econometrics},
+  volume={11},
+  number={3},
+  year={2007},
+  doi= {10.2202/1558-3708.1370},
[…]
