+++ testing/Simulation_Break_in_Persistence.Rmd
@@ -1,0 +1,372 @@
+---
+title: "Simulation: Break in Persistence"
+author: "Janis Becker"
+date: "`r Sys.Date()`"
+bibliography: paper.bib
+biblio-style: "apalike"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Simulation}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+To demonstrate that the implemented test do what they are supposed, we present simulations in the following which replicate those made by the authors of the tests. Here, we compare the critical values generated by the authors of the test with those when simulating our implemented test. To save computing time we consider a sample size of 100 observations and `M=10,000` replications. As `M` is typically chosen to be larger (50,000 or even higher) when simulating critical values we expect some deviations between our simulated critical values and those by the authors. These should, however, be rather small lying below 5 percent on average. 
+
+##Cusum test
+Following @leybourne2007cusum we perform the following simulation for the cusum test
+
+```{r, eval=FALSE,echo=TRUE}
+results_cusum<-vector(mode="list")
+set.seed(410)
+x=cumsum(rnorm(100))
+trend = c("none", "linear")
+for(a in 1:2){
+  tr=trend[a]
+  true=cusum_test(x,trend=tr)
+  simu=cusum_test(x,trend=tr,simu=1)
+  results_cusum[[a]]=true/simu
+}
+```
+
+This takes approximately thirty minutes and yields the following list of results depicting the ratio between our simulated critical values and those simulated by the authors
+
+```{r, echo = TRUE, eval=FALSE}
+> results_cusum
+[[1]]
+            90%       95%       99%
+Lower 0.9615506 0.9477438 0.9389883
+Upper 0.9988600 0.9899724 0.9300507
+
+[[2]]
+            90%      95%      99%
+Lower 1.0050942 1.011183 1.010665
+Upper 0.9995166 1.007026 1.018699
+```
+
+As can be seen, our simulated critical values and those by the authors coincide for the standard cusum test. Minor differences for very extreme quantile as the 99 percent quantile are due to the fact that our `M` is rather small. We should further note that we did not compare quantiles for the modified cusum test of @sibbertsen2009testing as these are based on T=10,000 so that there expected to be major differences when considering a small sample of T=100. 
+
+##LBI test
+Following @busetti2004tests we perform the following simulation for the LBI test
+
+```{r, echo = TRUE, eval=FALSE}
+results_LBI<-vector(mode="list")
+set.seed(410)
+x=rnorm(100)
+trend = c("none", "linear")
+statistic = c("mean", "max", "exp")
+q=1
+for(a in 1:2){
+  tr=trend[a]
+  for(b in 1:3){
+    st=statistic[b]
+    true=LBI_test(x,trend=tr,statistic=st)
+    simu=LBI_test(x,trend=tr,statistic=st,simu=1)
+    results_LBI[[q]]=true/simu
+    q=q+1
+  }
+}
+```
+
+This takes approximately five minutes and yields the following list of results depicting the ratio between our simulated critical values and those simulated by the authors
+
+```{r, echo = TRUE, eval=FALSE}
+> results_LBI
+[[1]]
+                                          90%      95%      99%
+Against change from I(0) to I(1)    0.9903047 0.996316 1.008091
+Against change from I(1) to I(0)    1.0398047 1.040524 1.084773
+Against change in unknown direction 0.9917469 1.022388 1.017279
+
+[[2]]
+                                          90%       95%       99%
[…]
