+++ R/pb_sim.R
@@ -6,8 +6,8 @@
 #' @param tp trend parameter, \code{t*tp} yields the contribution of the trend component if \code{trend="linear"}.
 #' @param d1 order of integration of the first part of the series.
 #' @param d2 order of integration of the second part of the series.
-#' @param mean mean of the series.
-#' @param var variance of the innovations.
+#' @param mean mean of the series. Default is \code{mean=0}.
+#' @param var variance of the innovations. Default is \code{var=1}
 #' @return Returns a vector containing the simulated time series.
 #' @author Janis Becker
 #' @examples
@@ -20,7 +20,7 @@
 #' stats::ts.plot(series)
 #' @export
 
-pb_sim<-function(T,tau,trend=c("none","linear"),tp=0,d1,d2,mean,var){
+pb_sim<-function(T,tau,trend=c("none","linear"),tp=0,d1,d2,mean=0,var=1){
   trend<-match.arg(trend,c("none","linear"))
   if(tau<=0 | tau>=1)
     stop("It must hold that 0<tau<1")
@@ -30,8 +30,8 @@ pb_sim<-function(T,tau,trend=c("none","linear"),tp=0,d1,d2,mean,var){
   T2<-round(T*(1-tau))
   if((T1+T2)>T){T2<-T2-1}
   if((T1+T2)<T){T2<-T2+1}
-  S1=LongMemoryTS::FI.sim(T1,q=1,rho=0,d=d1,var=var)+mean
-  S2=LongMemoryTS::FI.sim(T2,q=1,rho=0,d=d2,var=var)+S1[length(S1)]
+  S1=LongMemoryTS::FI.sim(T1,q=1,rho=0,d=d1,var=var,burnin=0)+mean
+  S2=LongMemoryTS::FI.sim(T2,q=1,rho=0,d=d2,var=var,burnin=0)+S1[length(S1)]
   series=c(S1,S2)
   if(trend=="linear") series<-series + 1:T*tp
   return(series)
