+++ README.md
@@ -29,11 +29,11 @@ install.packages("memochange")
 Examples
 --------
 
-In this section we present two short examples that illustrate how the implemented procedures can be used on a real data set. A more detailed presentation of the various tests and functions can be found in the vignette.
+In this section we present two short examples that illustrate how the implemented procedures can be used on a real data set. A more detailed presentation of the various tests and functions can be found in the vignettes.
 
 ### Tests for persistence change
 
-As an empirical example for a process that might exhibit a break in persistence, we consider the price of crude oil. First, we download the monthly price series from the FRED data base. For this purpose we need the `data.table` package.
+As an empirical example for a time series that might exhibit a break in persistence, we consider the price of crude oil. First, we download the monthly price series from the FRED data base. For this purpose we need the `data.table` package.
 
 ``` r
 oil=data.table::fread("https://fred.stlouisfed.org/graph/fredgraph.csv?bgcolor=%23e1e9f0&chart_type=line&drp=0&fo=open%20sans&graph_bgcolor=%23ffffff&height=450&mode=fred&recession_bars=on&txtcolor=%23444444&ts=12&tts=12&width=1168&nt=0&thu=0&trc=0&show_legend=yes&show_axis_titles=yes&show_tooltip=yes&id=MCOILWTICO&scale=left&cosd=1986-01-01&coed=2019-08-01&line_color=%234572a7&link_values=false&line_style=solid&mark_type=none&mw=3&lw=2&ost=-99999&oet=99999&mma=0&fml=a&fq=Monthly&fam=avg&fgst=lin&fgsnd=2009-06-01&line_index=1&transformation=lin&vintage_date=2019-09-23&revision_date=2019-09-23&nd=1986-01-01")
@@ -48,7 +48,7 @@ oil_xts=xts::xts(oil[,-1],order.by = oil$DATE)
 zoo::plot.zoo(oil_xts,xlab="",ylab="Price",main="Crude Oil Price: West Texas Intermediate")
 ```
 
-<img src="man/figures/README-unnamed-chunk-2-1.png" width="100%" style="display: block; margin: auto;" />
+<img src="man/figures/README-unnamed-chunk-3-1.png" width="100%" style="display: block; margin: auto;" />
 
 From the plot we observe that the series seems to be more variable in its second part from year 2000 onwards. This is first evidence that a change in persistence has occurred. We can test this hypothesis using the functions `cusum_test`, `LBI_test`, `LKSN_test`, `MR_test`, and `ratio_test`. In this short example we only consider the MR test as it is the most general one of the five implemented. The functionality of the other tests is similar. They all require a univariate numeric vector `x` as an input variable and yield a matrix of test statistic and critical values as an output variable.
 
@@ -61,19 +61,19 @@ Applying the default version of the MR test by Martins and Rodrigues (2014) yiel
 
 ``` r
 MR_test(x)
-#>                                           90%       95%       99%
-#> Against increase in memory          -1.638729 -1.925646 -2.503650
-#> Against decrease in memory          -1.653875 -1.953697 -2.509242
-#> Against change in unknown direction -1.937915 -2.210129 -2.726981
+#>                                          90%      95%      99%
+#> Against increase in memory          4.270666 5.395201 8.233674
+#> Against decrease in memory          4.060476 5.087265 7.719128
+#> Against change in unknown direction 5.065695 6.217554 9.136441
 #>                                     Teststatistic
-#> Against increase in memory              -3.300664
-#> Against decrease in memory              -1.629719
[…]
