+++ testing/Simulation_Break_in_Persistence.Rmd
@@ -18,7 +18,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-To demonstrate that the implemented test do what they are supposed, we present simulations in the following which replicate those made by the authors of the tests. Here, we compare the critical values generated by the authors of the test with those when simulating our implemented test. To save computing time we consider a sample size of 100 observations and `M=10,000` replications. As `M` is typically chosen to be larger (50,000 or even higher) when simulating critical values we expect some deviations between our simulated critical values and those by the authors. These should, however, be rather small lying below 5 percent on average. 
+The following simulations demonstrate that the implemented tests do what they are supposed. For this purpose, we simulate critical values based on our implemented tests and compare them to the critical values stated by the authors of the tests. To save computing time we consider a sample size of 100 observations and `M=10,000` replications. As `M` is typically chosen to be larger (50,000 or even higher) when simulating critical values we expect some deviations between our simulated critical values and those by the authors. These should, however, be rather small. 
 
 ##Cusum test
 Following @leybourne2007cusum we perform the following simulation for the cusum test
@@ -51,7 +51,7 @@ Lower 1.0050942 1.011183 1.010665
 Upper 0.9995166 1.007026 1.018699
 ```
 
-As can be seen, our simulated critical values and those by the authors coincide for the standard cusum test. Minor differences for very extreme quantile as the 99 percent quantile are due to the fact that our `M` is rather small. We should further note that we did not compare quantiles for the modified cusum test of @sibbertsen2009testing as these are based on T=10,000 so that there expected to be major differences when considering a small sample of T=100. 
+As can be seen, our simulated critical values and those by the authors coincide for the standard cusum test. Minor differences for very extreme quantile as the 99 percent quantile are due to the fact that our `M` is rather small. We should further note that we did not compare quantiles for the modified cusum test of @sibbertsen2009testing as these are based on T=10,000 so that we expect major differences when considering a small sample of T=100. 
 
 ##LBI test
 Following @busetti2004tests we perform the following simulation for the LBI test
