+++ man/fixbsupw.Rd
@@ -37,19 +37,25 @@ set.seed(410)
 # generate a fractionally integrated (long-memory) time series
 tseries  <- fracdiff::fracdiff.sim(n=T, d=d)$series
 
-# generate a fractionally integrated (long-memory) time series with a change in mean in the middle of the series
+# generate a fractionally integrated (long-memory) time series
+#  with a change in mean in the middle of the series
 changep  <- c(rep(0,T/2), rep(1,T/2))
 tseries2 <- tseries+changep
 
-# estimate the long-memory parameter of both series via local Whittle approach. The bandwidth to estimate d is chosen as T^0.65, which is usual in literature
+# estimate the long-memory parameter of both series via local 
+# Whittle approach. The bandwidth to estimate d is chosen 
+# as T^0.65, which is usual in literature
 d_est    <- LongMemoryTS::local.W(tseries, m=floor(1+T^0.65))$d
 d_est2   <- LongMemoryTS::local.W(tseries2, m=floor(1+T^0.65))$d
 
 # perform the test on both time series
 fixbsupw(tseries, d=d_est)
 fixbsupw(tseries2, d=d_est2)
-# For the series with no change in mean the test does not reject the null hypothesis of a constant mean across time at any reasonable significance level.
-# For the series with a change in mean the test rejects the null hypothesis at a 1\% significance level.
+# For the series with no change in mean the test does not reject 
+# the null hypothesis of a constant mean across time at any 
+# reasonable significance level.
+# For the series with a change in mean the test rejects the 
+# null hypothesis at a 1\% significance level.
 }
 \references{
 Iacone, F. and Leybourne, S. J. and Taylor, R. A. M. (2014): A fixed-b Test for a Break in Level at an unknown Time under Fractional Integration. Journal of Time Series Analysis, 35, pp. 40-54.
