+++ vignettes/Change_in_Mean.Rmd
@@ -1,0 +1,236 @@
+---
+title: "memochange-Tutorial: Change in Mean"
+author: "Kai Wenger"
+date: "`r Sys.Date()`"
+bibliography: paper.bib
+biblio-style: "apalike"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The `memochange` package can be used for two things: Checking for a break in persistence and checking for a change in mean. This vignette presents the functions related to a change in mean. This includes the functions `CUSUMfixed`, `CUSUMLM`, `CUSUM_simple`, `fixbsupw`, `snsupwald`, `snwilcoxon`, and `wilcoxonLM`. Before considering the usage of these functions, a brief literature review elaborates on their connection.
+
+## Literature Review
+
+In standard time series models it is usually assumed that the series have a constant mean over time. If this assumption is invalidated, inference and forecasts based on such models are misleading. Therefore, testing for a change in mean is of major importance.
+
+A typical example of a time series that could be subject to a change in mean is the series of average yearly temperatures in New Haven. It is visualized in the following graph.
+
+```{r, echo = TRUE, fig.height = 4, fig.width = 7, fig.align = "center"}
+utils::data(nhtemp)
+graphics::plot(nhtemp)
+```
+
+Three standard procedures to test for a change in mean at an unknown point in time are CUSUM tests originally proposed by @brown1975techniques, Wilcoxon-type rank tests (e.g. @bauer1972constructing), and sup-Wald tests by @andrews1993tests. Applying the standard CUSUM test, for example, we observe for the above temperature series that the corresponding p-value of the test is smaller than any resonable significance level. 
+
+```{r, echo = TRUE}
+strucchange::sctest(strucchange::efp(nhtemp ~ 1, type = "OLS-CUSUM"))
+```
+
+Therefore, it rejects the null hypothesis of a constant mean over time and we conclude that there is a change in mean.
+
+However, all these standard tests suffer from the issue that they cannot be applied under long memory. In a persistent long-memory time series far distant observations are significantly correlated. The degree of persistence is given by the long-memory parameter $d \in [0,0.5)$. Higher values of $d$ indicate higher persistence of the series. The special case $d=0$ is called short memory where far distant observations are not significantly correlated anymore. 
+The difference between a short-memory and a long-memory time series can be seen easily in the autocorrelation function (acf) of a series, which gives the correlation between observations separated by various time lags. 
+
+```{r, echo = TRUE, fig.height = 4, fig.width = 7, fig.align = "center"}
+T            <- 1000
+series_short <- fracdiff::fracdiff.sim(n=T,d=0)$series
+series_long  <- fracdiff::fracdiff.sim(n=T,d=0.45)$series
+
+graphics::par(mfrow=c(1,2))
+stats::acf(series_short,main="short memory")
+stats::acf(series_long,main="long memory")
+```
+
+We observe that the acf of the short-memory time series dies out quickly, while the autocorrelations in the long-memory time series are even at a very large lag (i.e. for far distant observations) high.
+
[…]
