+++ DESCRIPTION
@@ -28,7 +28,7 @@ Depends: R (>= 3.4.0)
 License: GPL-2
 Encoding: UTF-8
 LazyData: true
-Imports: urca (>= 1.3.0), forecast (>= 8.6), fracdiff (>= 1.4.2), LongMemoryTS (>= 0.1.0), sandwich (>= 2.5.1), strucchange (>= 1.5.1), longmemo (>= 1.1.1), stats (>= 3.6.1)
+Imports: urca (>= 1.3.0), forecast (>= 8.6), fracdiff (>= 1.4.2), LongMemoryTS (>= 0.1.0), sandwich (>= 2.5.1), strucchange (>= 1.5.1), longmemo (>= 1.1.1), stats (>= 3.4.1)
 RoxygenNote: 6.1.1
 Suggests: 
     testthat,
@@ -36,5 +36,7 @@ Suggests:
     rmarkdown,
     xts,
     zoo,
-    data.table
+    data.table,
+    utils,
+    graphics
 VignetteBuilder: knitr
