+++ R/CUSUMLM.R
@@ -57,6 +57,11 @@ CUSUMLM <- function(x,d,delta,tau=0.15)
     stop("x contains missing values")
   if(tau<=0 | tau>=1)
     stop("It must hold that 0<tau<1")
+  if(delta<0.6 | delta>0.8)
+    stop("It must hold that 0.6<delta<0.8")
+  if (mode(x) %in% ("numeric") == FALSE | is.vector(x) == 
+      FALSE) 
+    stop("x must be a univariate numeric vector")
   if(tau!=0.15)
     warning("Critical values are just implemented for tau=0.15")
   
