+++ README.Rmd
@@ -26,7 +26,7 @@ knitr::opts_chunk$set(
 
 An `R` package for identifying structural changes in time series. 
 
-# Description
+## Description
 
 For modeling and forecasting time series it is essential to know whether the series are stationary or non-stationary as many commonly applied statistical methods 
 such as OLS are invalid in case of non-stationarity. Examples of non-stationary series include processes that exhibit shifts in mean, i.e. the expected value of the series changes over time, 
@@ -35,7 +35,7 @@ The memochange package allows to consistently identify such changes in mean and 
 
 There is a wide range of macroeconomic and financial time series where the procedures can respectively should be applied. This includes beta, inflation rates, interest rates, trading volume, volatilities, and so on.
 
-# Installation
+## Installation
 
 You can install this R package from GitHub:
 
@@ -99,7 +99,7 @@ This can be estimated by the `BP_estim` function. It is important for the functi
 BP_estim(x,direction="01")
 ```
 
-This yields a list stating the location of the break (observation 151), semiparametric estimates of the order of integration in the two regimes (0.86 and 1.03) as well as the standard deviations of these estimates (0.13 and 0.15).
+This yields a list stating the location of the break (observation 151), semiparametric estimates of the order of integration (the persistence) in the two regimes (0.86 and 1.03) as well as the standard deviations of these estimates (0.13 and 0.15).
 
 ```{r, echo = TRUE}
 oil$DATE[151]
@@ -108,7 +108,7 @@ oil$DATE[151]
 Consequently, the function indicates that there is a break in persistence in July, 1998. 
 This means that from the beginning of the sample until June 1998 the series is integrated with an order of 0.85 and from July 1998 on the order of integration increased to 1.03.
 
-The function further allows for various types of break point and memory estimators. These are presented in the vignette. 
+The function further allows for various types of break point and persistence estimators. These are presented in the vignette. 
 
 ### Tests for change in mean
 
