+++ testing/Simulation_Change_in_Mean.Rmd
@@ -1,0 +1,523 @@
+---
+title: "Simulation: Change in Mean"
+author: "Kai Wenger"
+date: "`r Sys.Date()`"
+bibliography: paper.bib
+biblio-style: "apalike"
+output: rmarkdown::pdf_document
+vignette: >
+  %\VignetteIndexEntry{Simulation}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The following simulations demonstrate that the implemented tests for a change in mean in a long-memory time series do what they are supposed. For this purpose, we replicate results of papers that published the implemented tests. We use $500$ Monte Carlo replications throuhout all simulations.
+
+## Wilcoxon LM and self-normalized Wilcoxon test
+First, we replicate a comparision of the Wilcoxon LM test by @dehling2013non and the self-normalized Wilcoxon test by @betken2016testing, which is published in Tables III to VI of @betken2016testing pp. 793-794. For the power simulations (Tables V and VI) we just consider the case that the break is in the middle of the sample ($\tau=0.5$ following their notation) and the break magnitude is $\Delta=2$. Note that $d=H-0.5$.
+
+```{r, eval=FALSE,echo=TRUE}
+set.seed(410)
+# Parameter setting considered
+T_grid              <- c(10,50,100,500)
+d_grid              <- c(0.1,0.2,0.3,0.4)
+N                   <- 500
+
+# Generate array to save the results 
+resultmat           <- array(NA, dim=c(length(T_grid),length(d_grid),4))
+dimnames(resultmat) <- list(paste("T=",T_grid,sep=""),paste("d=",d_grid,sep=""),
+                            paste(rep(c("WilcoxonLM","snWilcoxon"),2),c("size","size",
+				"power, Delta=1, tau=0.5","power, Delta=1,tau=0.5"),sep=" "))
+
+# Monte Carlo simulation
+for(TTT in 1:length(T_grid))
+{
+  T <- T_grid[TTT]
+  for(ddd in 1:length(d_grid))
+  {
+    d                 <- d_grid[ddd]
+    result_vec        <- 0
+    for(i in 1:N)
+    {
+    # Simulate a fractionally integrated (long-memory) time series of 
+    # length T with memory d that is not subject to a shift.
+    tseries     <- fracdiff::fracdiff.sim(n=T,d=d)$series
+  
+    # Simulate a fractionally integrated (long-memory) time series of 
+    # length T with memory d that is subject to a shift in the middle of
+    # the sample of magnitude 2.  
+    changep     <- c(rep(0,T/2),rep(2,T/2))
+    tseries2    <- tseries+changep
+  
+    # They suppose that d is known and not estimated.
+    
+    # Apply both functions on both time series.
+    testIsize    <- wilcoxonLM(tseries,d=d)
+    testIIsize   <- snwilcoxon(tseries,d=d)
+    testIpower   <- wilcoxonLM(tseries2,d=d)
+    testIIpower  <- snwilcoxon(tseries2,d=d)
+    
+    # Save if the tests reject at the 5% significance level.
+    testIsize    <- testIsize["Teststatistic"]   > testIsize["95%"]
+    testIIsize   <- testIIsize["Teststatistic"]  > testIIsize["95%"]
[…]
