### 2019-11-11 KaiWenger/memochange#1 @KaiWenger
Hi @steffilazerte,
regarding this issue:

- we included a minimum version for each imported package in the `DESCRIPTION` file.
- we changed the `.Rbuildignore` file.
- we call functions directly by using the `package::function()` notation every time we apply functions from other packages.

Thank you for pointing these things out!
