### 2019-11-11 KaiWenger/memochange#2 @KaiWenger
Hi @steffilazerte,
regarding this issue:

- we included the .Rmd file
- we clarified the issues regarding the `README`.
- we include the output for all examples in the `README`.
- we added a simpler example based on a real data set to explain the change-in-mean tests. We further put the more complicated example into the new vignette.
