+++ paper/paper.md
@@ -31,9 +31,9 @@ For modeling and forecasting time series it is essential to know whether the ser
 
 ![Left: crude oil price over time, right: log squared returns of the NASDAQ index between 2006 and 2009.](plot.png)
 
-The left plot shows the monthly price of crude oil between 1986 and 2019. It can be seen that the series is more variable in the second part of the series from year 2000 onwards. This might be caused by a change in the autocovariance structure of the series. The right plot presents the daily log squared returns of the NASDAQ index between 2006 and 2009. It can be seen that in the second part of the series the mean of the log squared return seems to be larger than in the first part of the series. This is evidence for a change in mean, which is likely to be caused by the financial crisis started in August 2007. 
+The left plot shows the monthly price of crude oil between 1986 and 2019. It can be seen that the series is more variable in the second part of the series from year 2000 onwards. This might be caused by a change in the autocovariance structure of the series. The right plot presents the daily log squared returns of the NASDAQ index between 2006 and 2009. It can be seen that in the second part of the series the mean of the log squared return seems to be larger than in the first part of the series. This is evidence for a change in mean, which is likely to be caused by the financial crisis which started in August 2007. 
 
-The `memochange` package is an R (@RCoreTeam) package that allows to consistently identify such changes in mean and persistence. This helps to avoid model misspecification and improves forecasts of the time series. Concerning changes in persistence it implements the tests and breakpoint estimators by @leybourne2003tests, @busetti2004tests, @leybourne2004tests, @harvey2006modified, @leybourne2007cusum,  @sibbertsen2009testing, and @martins2014testing. Concerning change in mean it implements the tests by @horvath1997effect, @wang2008change, @shao2011simple, @dehling2013non, @iacone2014fixed, @betken2016testing, @wenger2018simple, and @wenger2019fixed. 
+The `memochange` package is an R (@RCoreTeam) package that allows to consistently identify such changes in mean and persistence. This helps to avoid model misspecification and improves forecasts of the time series. Concerning change in persistence it implements the tests and breakpoint estimators by @leybourne2003tests, @busetti2004tests, @leybourne2004tests, @harvey2006modified, @leybourne2007cusum,  @sibbertsen2009testing, and @martins2014testing. Concerning change in mean it implements the tests by @horvath1997effect, @wang2008change, @shao2011simple, @dehling2013non, @iacone2014fixed, @betken2016testing, @wenger2018simple, and @wenger2019fixed. 
 
 A related package is the `strucchange` package, which also allows to identify a change in mean. However, the procedures implemented here allow for valid inference in a more general setting. More detailed information can be found in the vignettes of the `memochange` package.
 
