### 2019-11-11 KaiWenger/memochange#3 @KaiWenger
Hi @steffilazerte,
regarding your issues:

#### General documentation
- we included two vignettes: one for the change-in-persistence tests and one for the change-in-mean tests, which explain in more detail the usage of the implemented functions.
- we changed the general description in the documentation for `memochange`.
#### Examples
- we comment all our examples step-by-step.
#### Unit Tests
- besides testing via travisCI, we include two types of additional tests: First, as you suggested, we include `testthat` tests for all of our functions. Second, we include the files `Simulation_Change_in_Mean` and `Simulation_Break_in_Persistence` in the testing folder where we reproduce results from published papers.
#### Community Guidelines
- we included a `CONTRIBUTING.md` file to make clear that software contributions are welcome.
