### 2019-11-11 KaiWenger/memochange#4 @KaiWenger
Hi @JonasMoss,
regarding this issue:

- we removed the `add_to_Namespace.R` file.
- we added the correct `lib::` prefix to all of the mentioned checkcomplaints.
