### 2019-11-11 KaiWenger/memochange#6 @KaiWenger
Hi @JonasMoss,
we did two types of additional tests:
1. We include `testthat` tests for all of our functions. As you suggested, the most reasonable combinations of use cases are tested. Furthermore, it is checked whether the implemeted tests produce reasonable results.
2. We reproduce results from many published papers in the files `testing\Simulation_Change_in_Mean` and `testing\Simulation_Break_in_Persistence`. Precisely, we reproduce Monte Carlo simulation studies as well as simulations of critical values of the papers that publish the tests we implement.

We also commented on this in [KaiWenger/memochange#3](https://github.com/KaiWenger/memochange/issues/3).
