### 2019-11-11 KaiWenger/memochange#7 @KaiWenger
Hi @JonasMoss,
we incorporated your remarks as follows:

- we documented our return values in a better way. We heavily thought about returning objects modeled after the functions of the `stats` package, but decided to skip this.
- we documented `x` as numeric time series and test for it to be a numeric time series in all of our functions.
- we changed the documentation about `tau`. We included an `error` message in all functions if `tau` is chosen in an incorrect manner. We further included a  `warning` message if the chosen `tau` is correct, but does not match the critical values deposited in the package.
- we (a) included short comments to our examples, (b) use the `set.seed` for reproducibility and (c) add `seealso{}` to our documentation.
- in the examples of the implemented functions we stayed with the simulated time series since then the user exactly knows how the tests should behave. However, we included an additional example in one of the two new vignettes and in the `README`. We further explained the real data examples in more detail there.
- we changed the arguments of part of the functions such that the name of the time series is always `x`.
- we named both `tau` since it is usual in literature. We hope that we changed the explanation in the arguments in a way such that no misunderstanding happens.
- we hope that we fixed all stylistic errors in the examples of all functions.

Once again, thank you very much for all these helpful and constructive suggestions!
