### 2019-11-11 KaiWenger/memochange#8 @KaiWenger
Hi @JonasMoss,
to not go into to much detail, let me summarize how we incorporated your remarks for all of our functions:

- we rewrote the error and warning messages of our functions to make them more clear.
- we changed all these things like writing `\code{"linear"}` instead of `\code{linear}`.
- as already mentioned, we call functions directly by using the `package::function()` notation.
- we added comments to all of our examples.
- we use `match.arg` for our checks.
- as already mentioned, we explained the arguments of all functions more precisely.
