### 2019-11-11 KaiWenger/memochange#9 @KaiWenger
Hi @JonasMoss,

we shortened the `Readme` and included an example for the change-in-mean tests on a real data set.
As you suggested, we further created two new vignettes - one for change-in-persistence tests and one for change-in-mean tests. In the vignettes we explain the problems that the implemented tests in much more detail than before.
