### 2019-11-11 #1820 @KaiWenger
Hi @steffilazerte and @JonasMoss,

first of all: thank you very much for your helpful and constructive thoughts and suggestions! We carefully considered all of your suggestions and think that our package improved significantly. We commented on all of your issues in [#1](https://github.com/KaiWenger/memochange/issues/1), [#2](https://github.com/KaiWenger/memochange/issues/2), [#3](https://github.com/KaiWenger/memochange/issues/3), [#4](https://github.com/KaiWenger/memochange/issues/4), [#5](https://github.com/KaiWenger/memochange/issues/5), [#6](https://github.com/KaiWenger/memochange/issues/6), [#7](https://github.com/KaiWenger/memochange/issues/7), [#8](https://github.com/KaiWenger/memochange/issues/8), and [#9](https://github.com/KaiWenger/memochange/issues/9). In summary, we incorporated your remarks as follows:

- we included a brief `Statement_of_Author_Contributions.md`.
- we incorporated the suggestions for "good-practices" / better documentation ([#1](https://github.com/KaiWenger/memochange/issues/1), [#4](https://github.com/KaiWenger/memochange/issues/4), [#5](https://github.com/KaiWenger/memochange/issues/5), [#7](https://github.com/KaiWenger/memochange/issues/7), [#8](https://github.com/KaiWenger/memochange/issues/8)). Among many things, this includes that
(a) we call functions directly by using the `package::function()` notation,
(b) we be more precise in explaining the arguments and return values of all functions,
(c) we explain the function's examples in more detail, and
(d) we changed the `DESCRIPTION` and `.Rbuildignore` files according to your suggestions.
- as both of you suggested, we included two types of additional tests:  First, we include automated `testthat` tests for all of our functions. We test for the most reasonable combinations of use cases and it is checked whether the implemeted tests produce reasonable results. Second, we reproduce trustable results from published scientific papers in the files `Simulation_Change_in_Mean` and `Simulation_Break_in_Persistence.` Precisely, we reproduce Monte Carlo simulation studies as well as simulations of critical values of the papers that published the tests we implement. We think that both tests increase the trust in our software significantly (see also [#3](https://github.com/KaiWenger/memochange/issues/3), [#6](https://github.com/KaiWenger/memochange/issues/6)).
- we included a `CONTRIBUTING.md` file to make clear that software contributions are welcome ([#3](https://github.com/KaiWenger/memochange/issues/3)).
- we made important changes in the `README`: we shortened it to be short and on the point, using little code and not much text. Furthermore, for the change-in-mean tests we include a simple real data example ([#2](https://github.com/KaiWenger/memochange/issues/2), [#3](https://github.com/KaiWenger/memochange/issues/3), [#9](https://github.com/KaiWenger/memochange/issues/9)).
- we created two new vignettes. One is about change-in-persistence tests. The other one is about change-in-mean tests. We put a lot of the content from the former `README` and `paper.md` in there, but also explained the functions and tests in much more detail and in many cases in an easier way ([#3](https://github.com/KaiWenger/memochange/issues/3), [#9](https://github.com/KaiWenger/memochange/issues/9)).
- we totally changed the `paper.md` to explain the functionality and purpose of our package "for a diverse, non-specialist audience" as stated in the review criteria of JOSS. As already mentioned, we use many parts of the former `paper.md` in our new vignettes. The new software paper is written in a much better way for a non-specialist audience. It further includes a comment on other packages for change in mean and change in persistence. We also elaborate on the statement of need at the end of the paper and shortened the paper. Since the link to our github repository is in the header of the software paper, we did not include a link once again at the bottom of the paper.

We hope that both of you are satisfied with our modifications. We really put a lot of effort the last weeks into the improvement of our package according to your suggestions. We are looking forward hearing from you :)
