### 2019-11-14 KaiWenger/memochange#4 @JonasMoss
Great! Thanks!
