### 2019-11-14 KaiWenger/memochange#7 @JonasMoss
The level of detail your documentation is outstanding. Thanks a lot for your work!
