### 2019-11-14 KaiWenger/memochange#6 @JonasMoss
Great work! Very nice coverage on your automated tests and the testing files is the icing on the cake and shows that this package is proved to do what it should to an unusually high degree.

Just a note though: Your tests can take long time to run and won't be accepted by CRAN I think. Consider changing the directory structure (following e.g. https://github.com/s-fleck/testthis) to make some tests run only when specifically triggered.
