### 2019-11-14 KaiWenger/memochange#9 @JonasMoss
Very nice!

One comment though: At least when I compile the vignette "change in mean" some of the citations fail to render properly.
