### 2019-11-14 #1820 @JonasMoss
@KaiWenger

Your response to our programming comments is great! I left two comments on the Github page that are non-critical.

I really like your new summary paper too. The only issue I can see is that you should cite the `strucchange` package. Add that and I can check off the last box.
