+++ vignettes/paper.bib
@@ -7,6 +7,7 @@
   year         = 2019,
   url          = {https://www.R-project.org}
 }
+
 @Article{dickey1981likelihood,
   Title={Likelihood ratio statistics for autoregressive time series with a unit root},
   Author={Dickey, David A and Fuller, Wayne A},
@@ -16,6 +17,39 @@
   doi={10.2307/1912517},
   Publisher={JSTOR}
 }
+
+@Article{andrews1993tests,
+  author    = {Andrews, Donald WK},
+  title     = {Tests for parameter instability and structural change with unknown change point},
+  journal   = {Econometrica},
+  year      = {1993},
+  pages     = {821--856},
+  doi       = {10.2307/2951764},
+  publisher = {JSTOR},
+}
+
+@article{bauer1972constructing,
+  title={Constructing confidence sets using rank statistics},
+  author={Bauer, David F},
+  journal={Journal of the American Statistical Association},
+  volume={67},
+  number={339},
+  pages={687--690},
+  year={1972},
+  doi       = {10.1080/01621459.1972.10481279},
+  publisher={Taylor \& Francis}
+}
+
+@article{brown1975techniques,
+  title={Techniques for testing the constancy of regression relationships over time},
+  author={Brown, Robert L and Durbin, James and Evans, James M},
+  journal={Journal of the Royal Statistical Society. Series B (Methodological)},
+  pages={149--192},
+  year={1975},
+  doi       = {10.1111/j.2517-6161.1975.tb01532.x},
+  publisher={JSTOR}
+}
+
 
 @Article{kwiatkowski1992testing,
   Title={Testing the null hypothesis of stationarity against the alternative of a unit root: How sure are we that economic time series have a unit root?},
@@ -233,7 +267,7 @@
 
 @article{robinson1995gaussian,
   title={Gaussian semiparametric estimation of long range dependence},
-  author={Robinson, Peter M and others},
+  author={Robinson, Peter M},
   journal={The Annals of Statistics},
   volume={23},
   number={5},
@@ -301,6 +335,31 @@
   publisher={Springer}
 }
 
+@Article{kramer2000testing,
+  author  = {Kr{\"a}mer, Walter and Sibbertsen, Philipp},
+  title   = {Testing for Structural Changes in the Presence of Long Memory},
+  journal = {International Journal of Business and Economics},
+  year    = {2002},
+  volume  = {1},
+  number  = {3},
+  pages   = {235--242},
+  url     = {http://hdl.handle.net/10419/77279},
+}
+
+
+@Article{wright1998testing,
+  author    = {Wright, Jonathan H},
+  title     = {Testing for a Structural Break at Unknown Date with Long-memory Disturbances},
+  journal   = {Journal of Time Series Analysis},
+  year      = {1998},
+  volume    = {19},
+  number    = {3},
+  pages     = {369--376},
+  doi       = {10.1111/1467-9892.00097},
+  publisher = {Wiley Online Library},
+}
+
+
 @article{Xu14,
   title={Forecasting return volatility: Level shifts with varying jump probability and mean reversion},
   author={Xu, Jiawen and Perron, Pierre},
@@ -322,5 +381,50 @@
   pages={1714--1726},
   year={2011},
   doi = {10.1016/j.jbankfin.2010.11.007},
+  publisher={Elsevier}
+}
+
+@article{robinson2005robust,
+  title={Robust covariance matrix estimation: HAC estimates with long memory/antipersistence correction},
+  author={Robinson, Peter M},
+  journal={Econometric Theory},
+  volume={21},
+  number={01},
+  pages={171--180},
+  year={2005},
[…]
