### 2019-11-14 KaiWenger/memochange#9 @Janis410
It appears we forgot to update the .bib file on github.  Corrected that mistake, so all citations should render properly now.
Thanks for bringing that up!
