+++ paper/paper.bib
@@ -324,3 +324,16 @@
   doi = {10.1016/j.jbankfin.2010.11.007},
   publisher={Elsevier}
 }
+
+  @Article{strucchange,
+    title = {strucchange: An R Package for Testing for Structural
+      Change in Linear Regression Models},
+    author = {Achim Zeileis and Friedrich Leisch and Kurt Hornik and
+      Christian Kleiber},
+    journal = {Journal of Statistical Software},
+    year = {2002},
+    volume = {7},
+    number = {2},
+    pages = {1--38},
+    url = {http://www.jstatsoft.org/v07/i02/},
+  }
