+++ paper/paper.bib
@@ -327,7 +327,7 @@
 
 
 @Manual{strucchange,
-    title={strucchange: An R Package for Testing for Structural},
+    title={strucchange: An R Package for Testing for Structural Change in Linear Regression Models},
     author={Achim Zeileis and Friedrich Leisch and Kurt Hornik and Christian Kleiber},
     year={2019},
     note={R package version 1.5.1},
