### 2019-11-14 #1820 @JonasMoss
@KaiWenger Please use the Journal of Statistical Software entry from https://cran.r-project.org/web/packages/strucchange/citation.html
