### 2019-11-14 KaiWenger/memochange#6 @Janis410
Thanks. We modified the tests using skip_on_cran() to reduce the testing duration to 14 seconds on CRAN. Hope they are fine with that.
We also updated the change in mean vignette such that the simulation is not actually run in R which decreases the time to build the vignette.
