+++ paper/paper.bib
@@ -325,11 +325,15 @@
   publisher={Elsevier}
 }
 
-
-@Manual{strucchange,
-    title={strucchange: An R Package for Testing for Structural Change in Linear Regression Models},
-    author={Achim Zeileis and Friedrich Leisch and Kurt Hornik and Christian Kleiber},
-    year={2019},
-    note={R package version 1.5.1},
-    url={https://cran.r-project.org/package=strucchange},
-}
+  @Article{strucchange,
+    title = {strucchange: An R Package for Testing for Structural
+      Change in Linear Regression Models},
+    author = {Achim Zeileis and Friedrich Leisch and Kurt Hornik and
+      Christian Kleiber},
+    journal = {Journal of Statistical Software},
+    year = {2002},
+    volume = {7},
+    number = {2},
+    pages = {1--38},
+    url = {http://www.jstatsoft.org/v07/i02/},
+  }
