### 2019-11-14 #1820 @KaiWenger
@JonasMoss Thanks for your comments! We included the citation for the strucchange package in our software paper and treated your last two comments on our Github page :)
