### 2019-11-14 #1820 @JonasMoss
Great! Just add the DOI to the paper. =) The DOI is 10.18637/jss.v007.i02
