+++ paper/paper.bib
@@ -325,7 +325,7 @@
   publisher={Elsevier}
 }
 
-  @Article{strucchange,
+@Article{strucchange,
     title = {strucchange: An R Package for Testing for Structural
       Change in Linear Regression Models},
     author = {Achim Zeileis and Friedrich Leisch and Kurt Hornik and
@@ -335,5 +335,6 @@
     volume = {7},
     number = {2},
     pages = {1--38},
+    doi={10.18637/jss.v007.i02},
     url = {http://www.jstatsoft.org/v07/i02/},
   }
