### 2019-11-14 #1820 @steffilazerte
Hi @KaiWenger, @Janis410,
Wow! So much work in such a short time. Adding the tests, vignettes and annotating the examples makes this a very solid package.

I also like the re-write of the article, it's much more accessible and is very clear about it's purpose.

I have a couple of editorial (and occasionally nit-picky) comments for the article and one comment for the repository:
- [ ] The text always refers to "mean" then "persistence" (in that order). To maintain parallelism, could the figure be reversed (i.e. mean on the left, persistence on the right) and the example text made to follow the same order? That way the reader always reads them in the same order and it makes it easier to follow
- [ ] P1 Paragraph 3 "The memochange package is an R (R Core Team, 2019) package that **identifies** such changes in mean and persistence."
- [ ] P2 paragraph 2 "A related package is the `strucchange` package (ref), which also **identifies changes** in mean. However, the procedures implemented in the **memochange package** allow for valid inference in a more general setting."

My final suggestion is up to you:
- as @JonasMoss suggested in KaiWenger/memochange#9, it would be a good idea to include a description in your GitHub repository (i.e. Replace the text right at the top of the page where it says "No description, website, or topics provided.")

A really great solid package, thanks!
