### 2019-11-15 #1820 @KaiWenger
Hi @steffilazerte, Hi @JonasMoss,

thanks again for your reviews and nice words :)
We
- added the DOI to the strucchange package
- changed the picture and the example in the software paper to maintain the parallelism of "mean" and "persistence"
- corrected the typos in the software paper
- added a description to the GitHub repository (sorry, we totally overlooked this in our first revision!)
