### 2019-11-15 #1820 @KaiWenger
Hi @mbobra,
the boxes in [#1820 (comment)](https://github.com/openjournals/joss-reviews/issues/1820#issuecomment-547599162) are all checked (even though we did not do it) and we completed our work on all remaining small changes that were suggested :)
