### 2019-11-15 #1820 @steffilazerte
Hi @KaiWenger, @Janis410 ,
So close! Just the text below the figure should be reversed (it doesn't match the figure), and I would suggest explicitly including the word "persistence" to help readers who are unfamiliar.

Now it is:
> The left plot shows the monthly price of crude oil between 1986 and 2019. It can be seen
that the series is more variable in the second part from year 2000 onwards. This might be
caused by a change in the autocovariance structure of the series. The right plot presents the
daily log squared returns of the NASDAQ index between 2006 and 2009. It can be seen that
in the second part of the series the mean of the log squared return seems to be larger than in
the first part of the series. This is evidence for a change in mean, which is likely to be caused
by the financial crisis started in August 2007.

But it should be:

The **left** plot presents the daily log squared returns of the NASDAQ index between 2006 and 2009. It can be seen that in the second part of the series the mean of the log squared return seems to be larger than in the first part of the series. This is evidence for a change in mean, which is likely to be caused by the financial crisis started in August 2007. The **right** plot shows the monthly price of crude oil between 1986 and 2019. It can be seen that the series is more variable in the second part from year 2000 onwards. This might be caused by a change in the autocovariance structure **(i.e. a change in persistence)** of the series.

(I checked off the boxes in the[ above comment](https://github.com/openjournals/joss-reviews/issues/1820#issuecomment-547599162) as I evaluated your responses to them :smile:)
