### 2019-11-15 #1820 @KaiWenger
@mbobra : our last change of the software paper was to include @steffilazerte and @JonasMoss in our acknowledgements :smiley:
We feel that our package improved significantly by including their comments and suggestions and we would be glad if the the package is now suitable for publication in the Journal of Open Source Software :relaxed:
