### 2019-11-15 #1820 @mbobra
👋 @JonasMoss @steffilazerte Thank you both so much for your incredibly thorough reviews. Yours are some of the most detailed reviews I've ever seen. I sincerely appreciate all your time and effort. JOSS is lucky to have reviewers like you ☀️
