### 2019-11-15 #1820 @mbobra
👋 Great news, @KaiWenger 😄 We're almost there! Can you please create a release and [archive it on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI? Can you please add the DOI in your README.md file, too? After that I think we're done 🎉
