+++ README.md
@@ -1,26 +1,41 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-memochange
-==========
 
-Testing for Structural Breaks under Long Memory and Testing for Changes in Persistence
---------------------------------------------------------------------------------------
+# memochange
 
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/memochange)](https://CRAN.R-project.org/package=memochange) [![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange) [![License: GPL v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0) [![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+## Testing for Structural Breaks under Long Memory and Testing for Changes in Persistence
+
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/memochange)](https://CRAN.R-project.org/package=memochange)
+[![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange)
+[![License: GPL
+v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0)
+[![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+[![DOI](https://zenodo.org/badge/210588375.svg)](https://zenodo.org/badge/latestdoi/210588375)
 
 An `R` package for identifying structural changes in time series.
 
-Description
------------
+## Description
 
-For modeling and forecasting time series it is essential to know whether the series are stationary or non-stationary since many commonly applied statistical methods (such as OLS) are invalid under non-stationarity. Two features that cause a time series to be non-stationary are considered here. On the one hand a time series can be subject to a change in mean, i.e. the expected value of the series changes over time. On the other hand a time series can be subject to a break in the autocovariance often refered to as a change in persistence, i.e. the dependence structure of the series changes over time.
+For modeling and forecasting time series it is essential to know whether
+the series are stationary or non-stationary since many commonly applied
+statistical methods (such as OLS) are invalid under non-stationarity.
+Two features that cause a time series to be non-stationary are
+considered here. On the one hand a time series can be subject to a
+change in mean, i.e.  the expected value of the series changes over
+time. On the other hand a time series can be subject to a break in the
+autocovariance often refered to as a change in persistence, i.e.  the
+dependence structure of the series changes over time.
 
-The memochange package allows to consistently identify such changes in mean and persistence. This helps to avoid model misspecification and improve forecasting the series.
+The memochange package allows to consistently identify such changes in
+mean and persistence. This helps to avoid model misspecification and
+improve forecasting the series.
 
-Potential examples for series with a change in mean and/or in persistence are found in the macroeconomic and financial area. This includes beta, inflation rates, interest rates, trading volume, volatilities, and so on.
+Potential examples for series with a change in mean and/or in
+persistence are found in the macroeconomic and financial area. This
+includes beta, inflation rates, interest rates, trading volume,
+volatilities, and so on.
 
-Installation
-------------
+## Installation
 
 You can install this R package from GitHub:
 
@@ -36,14 +51,19 @@ or directly from the CRAN repository:
 install.packages("memochange")
 ```
 
-Examples
---------
+## Examples
 
[…]
