+++ README.Rmd
@@ -23,6 +23,7 @@ knitr::opts_chunk$set(
 [![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/memochange)](https://CRAN.R-project.org/package=memochange)
 [![License: GPL v2](https://img.shields.io/badge/License-GPLv2-blue.svg)](https://www.gnu.org/licenses/gpl-2.0) 
 [![BuildStatus](https://api.travis-ci.org/KaiWenger/memochange.svg?branch=master)](https://travis-ci.com/KaiWenger/memochange)
+[![DOI](https://zenodo.org/badge/210588375.svg)](https://zenodo.org/badge/latestdoi/210588375)
 
 An `R` package for identifying structural changes in time series. 
 
