### 2019-11-16 #1820 @KaiWenger
Hi @mbobra :wave:
we created a release (for the release we changed the version from v1.0.1 to v1.1.0), archieved it on Zenodo, obtained the DOI and added it in our `README.md` :smiley:
