### 2019-11-18 #1820 @mbobra
@openjournals/joss-eics This paper is ready for acceptance! Nice work @KaiWenger 🎉
