### 2019-11-19 #1820 @kthyng
I fixed a typo in the paper if you want to use the patch: https://github.com/KaiWenger/memochange/compare/master...kthyng:patch-1
