### 2019-11-20 #1820 @KaiWenger
Hi @kthyng :smiley:
We fixed the typo and updated our metadata at Zenodo :blush:
Thanks for pointing that out!
